fix: tag value handles with the runtime that issued them
Two runtimes hand out bit-identical handles -- the first value each
issues is index 0, generation 1 -- so resolving one runtime's handle
against the other succeeded and answered with an unrelated value of the
same shape. Confirmed directly before changing anything: with both
registries populated, A's handle read through B returned BK_OK and B's
string.

The existing multi-runtime test appeared to cover this, but it only
evaluated in A. B's registry was empty, so every index was out of range
and the rejection was incidental rather than a property of the handle.
The test now populates both sides first, which is what makes it a real
check.

bk_value widens to 64 bits and carries the issuing runtime's id in the
top 16:

    63..48        47        46..32          31..0
   [ rt id ]   [scope]  [ generation ]   [ index+1 ]

A mismatch is now BK_ERR_INVALID. Scope handles carry the id too, so an
argument handle from one runtime's callback cannot be resolved by
another. The id comes from a process-wide counter that wraps at 16 bits,
so a collision needs 65536 runtimes opened and a handle from an old one
still held -- a far narrower window than having no id at all.

The index field grows from 16 to 32 bits as a side effect, so the 65535
live-handle ceiling is gone; the registry is bounded by memory now.

bindings/c/two_runtimes.c existed largely to demonstrate the old
footgun, printing "D's value, not C's" and calling the caught case luck.
Rewritten to show the guarantee instead.

Verified: test/run_local.sh green, all five C ABI suites pass, both C3
examples correct.

+++ include/boomkat.h
@@ -49,6 +49,7 @@
 #define BOOMKAT_H
 
 #include <stddef.h>
+#include <stdint.h>
 
 #ifdef __cplusplus
 extern "C" {
@@ -63,8 +64,15 @@ extern "C" {
 /* Opaque runtime handle. */
 typedef void *bk_runtime;
 
-/* Opaque value handle. 0 is never a valid handle. */
-typedef unsigned int bk_value;
+/*
+ * Opaque value handle. 0 is never a valid handle.
+ *
+ * 64 bits, not a pointer: the low half names a slot in the issuing runtime's
+ * GC-rooted registry, and the high half carries that runtime's id, so a handle
+ * used against the wrong runtime is rejected rather than silently resolving to
+ * an unrelated value of the same shape.
+ */
+typedef uint64_t bk_value;
 #define BK_INVALID_VALUE ((bk_value)0)
 
 /*
